s3: Fix the GNU ld version detection on SLES
On SLES 11 "ld -v" gives

GNU ld (GNU Binutils; SUSE Linux Enterprise 11) 2.19

The regexp to intended to detect the "2.19" is confused by the "11" in the
version string. I'm not really into regexps, and awk '{print $NF}' to me is a
simple (and hopefully portable) way to print the last field.