Revert "socket-wrapper: not all systems have FIONREAD defined"
This reverts commit 710aa773d54509de34404f9992c5058ddfa45f3b.

We rely on FIONREAD in a lot of other parts in the code,
so there's no need to have an ifdef for it in the socket_wrapper code.

On tru64 FIONREAD is defined in <sys/ioctl.h>
and we include <sys/ioctl.h> via "system/network.h".

Tridge: maybe a HAVE_SYS_IOCTL_H was missing at the time you tried
it on tru64?

If we find a platform that doesn't support it,
we need to bail out at configure time or
provide a replacement in libreplace.

metze