s4:WAF buildsystem - support out of "source4" directory builds
Bug: https://bugzilla.samba.org/show_bug.cgi?id=4081

With this patch we are able to invoke s4 builds from the outside of the
"source4" directory (but the target remains the "source4/bin" path).

One constraint: all commands: "autogen-waf.sh", "configure", "make" have to be
run from the same directory!
Regarding "make": you have to run it using "make -C <source4 path> [targets]" if
the invoke directory is not "source4" itself.