tdb: use tdb_nest_lock() for open lock.
This never nests, so it's overkill, but it centralizes the locking into
lock.c and removes the ugly flag in the transaction code to track whether
we have the lock or not.

Note that we have a temporary hack so this places a real lock, despite
the fact that we are in a transaction.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>

(Imported from commit 9136818df30c7179e1cffa18201cdfc990ebd7b7)

(This used to be ctdb commit 6812d81907793299e874f121174d885f6500f374)