tdb: cleanup: always grab allrecord lock to infinity.
We were previously inconsistent with our "global" lock: the
transaction code grabbed it from FREELIST_TOP to end of file, and the
rest of the code grabbed it from FREELIST_TOP to end of the hash
chains.  Change it to always grab to end of file for simplicity and
so we can merge the two.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>

(Imported from commit 9341f230f8968b4b18e451d15dda5ccbe7787768)

(This used to be ctdb commit 46f2c33357c999c31a8064c159c6162269c28d9d)