Revert "Make -k a simple non-bool option."
This reverts commit 2bbf055dbb15ea99b80ae66adc906ad5494a3999.

This change broke 'make test', because the current syntax is deliberate
and an integral part of how Samba4 uses Kerberos.

If a change like this is to be implemented, then careful review of all
our test scripts must be taken first, and then new options such as
--dont-use-kerberos, --may-use-kerberos and --must-use-kerberos
should be added.

Andrew Bartlett