s4/test: make test_cleanup() in libnet_user library more robust
test_cleanup() is called always with RDN name of the user to be deleted.
When modify-user test fails however, we end up with a user
with RDN = libnetusertest and samAccountName = random_name.
This way we can not delete the user and the error message is
quite misleading (I've spent a *lot* of time trying to figure out
if the database is corrupted because of this error).