s3: set pthread cflags in a more portable fashion
this avoids one place of a GNU make dependency.

Portable make sucks a bit as it has no way of defining variables for a limited
set of targets only. Manual duplication of target rules is the only way to
achieve what we need.

If someone actually knows a trick how to do this without target duplication,
let me know.